DRF_stocks_products/requirements.txt
django~=4.1.5
djangorestframework~=3.14.0
psycopg2-binary
flake8
python-dotenv~=0.21.1
django-filter~=22.1
gunicorn
expects